feat: refactor AI agent utilities and enhance connection secret management
